extract base webtest
